pyproject.toml
[project]
name = "columbina"
version = "1.4.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0.0,<2.1.0",
    "sqlmodel==0.0.22",
    "psycopg2-binary>=2.9.0",
    "openai>=1.101.0",
    "pillow>=11.3.0",
    "py-cord",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "mem0ai>=1.0.0",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
]

[tool.uv.sources.py-cord]
git = "https://github.com/Pycord-Development/pycord.git"
rev = "v2.7.x"
